floats

+++ src/main/java/db/DirectIndexCollectionBridge.java
@@ -8,12 +8,12 @@ public class DirectIndexCollectionBridge {
     public static final String DocumentKey = "docName";
     public static final String DocumentValue = "values";
 
-    public static void addDirectIndexes(HashMap<String, HashMap<String, Integer>> directIndexes) {
-        Iterator<Map.Entry<String, HashMap<String, Integer>>> iterator = directIndexes.entrySet().iterator();
+    public static void addDirectIndexes(HashMap<String, HashMap<String, Float>> directIndexes) {
+        Iterator<Map.Entry<String, HashMap<String, Float>>> iterator = directIndexes.entrySet().iterator();
         List<Document> documents = new ArrayList<>();
 
         while (iterator.hasNext()) {
-            Map.Entry<String, HashMap<String, Integer>> next = iterator.next();
+            Map.Entry<String, HashMap<String, Float>> next = iterator.next();
             Document toInsert = new Document()
                     .append(DocumentKey, next.getKey())
                     .append(DocumentValue, next.getValue());
@@ -27,7 +27,7 @@ public class DirectIndexCollectionBridge {
         }
     }
 
-    public static void addDirectIndex(String key, HashMap<String, Integer> value) {
+    public static void addDirectIndex(String key, HashMap<String, Float> value) {
         Document toInsert = new Document()
                 .append(DocumentKey, key)
                 .append(DocumentValue, value);
